[Refactor] : login repository domain 레이어로 이동
Git : #1
