+++ lib/_testt/presentation/home/cards/transport_card.dart
@@ -1,5 +1,5 @@
 import 'dart:ui';
-import 'package:earlybuddy/_testt/presentation/asset/EBFont.dart';
+import 'package:earlybuddy/shared/assets/EBFont.dart';
 import 'package:flutter/material.dart';
 
 class TransportCard extends StatelessWidget {
