+++ lib/domain/auth_repository/auth_repository.dart
@@ -1,0 +1,1 @@
+export 'source/auth_repository.dart';
