+++ lib/domain/user_repository/source/model/model.dart
@@ -1,0 +1,1 @@
+export 'user.dart';
