+++ lib/domain/user_repository/user_repository.dart
@@ -1,0 +1,2 @@
+export 'source/user_repository.dart';
+export 'source/model/model.dart';
