+++ lib/presentation/auth/bloc/auth_bloc.dart
@@ -1,10 +1,8 @@
 import 'dart:async';
-import 'package:earlybuddy/login/repository/auth_repository/auth_repository.dart';
-import 'package:earlybuddy/login/repository/user_repository/user_repository.dart';
 import 'package:equatable/equatable.dart';
-import 'package:earlybuddy/login/model/model.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-
+import 'package:earlybuddy/domain/auth_repository/auth_repository.dart';
+import 'package:earlybuddy/domain/user_repository/user_repository.dart';
 part 'auth_event.dart';
 part 'auth_state.dart';
