+++ lib/presentation/auth/bloc/auth_event.dart
