+++ lib/presentation/auth/bloc/auth_state.dart
