+++ lib/presentation/home/view/home_view.dart
@@ -1,4 +1,4 @@
-import 'package:earlybuddy/auth/auth.dart';
+import 'package:earlybuddy/presentation/auth/auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
