+++ lib/presentation/login/model/model.dart
@@ -1,3 +1,2 @@
 export 'password.dart';
 export 'email.dart';
-export 'user.dart';
