+++ lib/presentation/login/view/login_view.dart
@@ -1,5 +1,5 @@
-import 'package:earlybuddy/login/login.dart';
-import 'package:earlybuddy/login/repository/auth_repository/auth_repository.dart';
+import 'package:earlybuddy/domain/auth_repository/auth_repository.dart';
+import 'package:earlybuddy/presentation/login/login.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
