+++ lib/presentation/splash/splash.dart
@@ -1,0 +1,1 @@
+export 'view/view.dart';
