+++ lib/presentation/splash/view/view.dart
@@ -1,0 +1,1 @@
+export 'splash_view.dart';
